Contract and trait for resolving values
